pyproject.toml
[project]
name = "ghidra-web-analyzer"
version = "1.0.0"
description = "Flask-based web application for Ghidra binary analysis with custom MCP integration"
authors = [
    { name = "ShadowSeek Team", email = "dev@shadowseek.security" }
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Flask Web Application
    "flask>=2.3.3",
    "flask-sqlalchemy>=3.0.5",
    "flask-migrate>=4.0.5",
    "flask-cors>=4.0.0",
    "flask-socketio>=5.3.6",
    # Async Web Server
    "aiohttp>=3.9.1",
    "aiohttp-cors>=0.7.0",
    "websockets>=12.0",
    # MCP Protocol
    "jsonrpc-base>=2.2.0",
    "jsonrpc-websocket>=3.1.5",
    "jsonrpc-async>=2.1.2",
    # Database
    "sqlalchemy>=2.0.23",
    # Task Queue
    "celery>=5.3.4",
    "kombu>=5.3.3",
    # File Handling
    "werkzeug>=2.3.7",
    "python-magic>=0.4.27",
    # System Utilities
    "python-dotenv>=1.0.0",
    "psutil>=5.9.6",
    "pathlib2>=2.3.7.post1",
    # HTTP Requests
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "openai>=1.90.0",
    "flask-restx==1.3.0",
    "setuptools>=80.9.0",
    "ghidra-bridge>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flask_app", "mcp_server"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 
